Remove code duplication by moving common functionality into a new BaseRootController stuff.
